# ApparelPro.WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApparelPro.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/address/list": {
      "get": {
        "tags": [
          "Address"
        ],
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sortColumn",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterColumn",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterQuery",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModelPaginationAPIModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModelPaginationAPIModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModelPaginationAPIModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/address/list/AddressId/{addressId}": {
      "get": {
        "tags": [
          "Address"
        ],
        "operationId": "GetAddressesByAddressIdAsync",
        "parameters": [
          {
            "name": "addressId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressAPIModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressAPIModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressAPIModel"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/address/list/byAddressId": {
      "get": {
        "tags": [
          "Address"
        ],
        "operationId": "GetAddressesForBuyerByAddressIdAsync",
        "parameters": [
          {
            "name": "addressId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sortColumn",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterColumn",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterQuery",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModelPaginationAPIModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModelPaginationAPIModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModelPaginationAPIModel"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/address/list/byIdAndAddressId": {
      "get": {
        "tags": [
          "Address"
        ],
        "operationId": "GetAddressByIdAndAddresIdAsync",
        "parameters": [
          {
            "name": "addressId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressAPIModel"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/address": {
      "post": {
        "tags": [
          "Address"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressAPIModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressAPIModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressAPIModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "tags": [
          "Address"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "addressId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressAPIModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressAPIModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressAPIModel"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/address/{id}/{addressId}": {
      "delete": {
        "tags": [
          "Address"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "addressId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressAPIModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "addressId": {
            "type": "string",
            "format": "uuid"
          },
          "addressType": {
            "$ref": "#/components/schemas/AddressType"
          },
          "streetAddress": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "postCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "default": {
            "type": "boolean",
            "nullable": true
          },
          "buyerCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddressAPIModelPaginationAPIModel": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressAPIModel"
            },
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "sortColumn": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "string",
            "nullable": true
          },
          "filterColumn": {
            "type": "string",
            "nullable": true
          },
          "filterQuery": {
            "type": "string",
            "nullable": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "AddressType": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "CreateAddressAPIModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "addressId": {
            "type": "string",
            "format": "uuid"
          },
          "addressType": {
            "$ref": "#/components/schemas/AddressType"
          },
          "streetAddress": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "postCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "default": {
            "type": "boolean",
            "default": false,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UnprocessableEntityResult": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "UpdateAddressAPIModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "addressId": {
            "type": "string",
            "format": "uuid"
          },
          "addressType": {
            "$ref": "#/components/schemas/AddressType"
          },
          "streetAddress": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "postCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "default": {
            "type": "boolean",
            "nullable": true
          },
          "buyerCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter token to login to ApparelPro",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}